+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.5] - 2023-04-14
+
+### Added
+
+- Wifi reconnection option after connecting to ethernet
+
+### Fixed
+
+- Ethernet not connecting due to attempt to connect to same network as Wireless LAN
+
 ## [0.0.4] - 2023-03-25
 
 ### Added